requirements.txt
discord.py
pytz
pyrebase4